Add files via upload
Since 2016, almost 60,000 new leukaemia cases have been found in only the United States of America. Furthermore, it suggests that leukaemia is the cancer kind that most usually strikes people under the age of twenty. Even if the study is based on a Western country, Bangladesh, an Asian country with a substandard healthcare system, should be concerned about it. According to research, 83% or so people with chronic lymphocytic leukaemia survive the disease for five years. This research focuses on acute lymphocytic leukaemia (ALL), which is the type of leukaemia that is most common in Bangladesh. Oncologists are aware that cancer can often be treated successfully if it is discovered in the early stages.The treatment must therefore begin as soon as is practical. We suggest a hands-on method for identifying abnormal blood components, such as neutrophils, eosinophils, basophils, lymphocytes, and monocytes, which are frequently observed in cancer patients. In the present study, we first identify 14 qualities to prepare the dataset and then discover 4 important attributes that significantly contribute to the identification of a leukaemia patient. Additionally, 256 primary data from leukaemia patients were gathered by us. The data is then analysed using a microscope to obtain images, which are then fetched into a machine learning algorithm called Faster-RCNN to determine the likelihood that cancer cells would emerge. In order to identify the comparable blood item, we have implemented two loss functions to the classifier model and the RPN (Region Convolutional Neural Network) model.The excessive synthesis of abnormal white platelets leads to leukaemia.